Update power supply
Update power supply classes:
 - create interface for lan exchanger with base exchange methods;
 - division into subsystem for power supply;
 - created an initialization file for working with the library;
 - create for all subsystem interfaces and classes;
 - renamed classes.
