+++ Devices/PowerSupply/AgilentN5746A/N5746A.cs
@@ -1,12 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Security.Policy;
-using System.Text;
-using System.Threading.Tasks;
+using DevicesControlLibrary.Devices.PowerSupply.Subsystems.Calibration;
+using DevicesControlLibrary.Devices.PowerSupply.Subsystems.Common;
 using DevicesControlLibrary.Exchange;
 
-namespace DevicesControlLibrary.Devices.PowerSupply
+namespace DevicesControlLibrary.Devices.PowerSupply.AgilentN5746A
 {
     /// <summary>
     ///     Work with Power Supply Agilent / Keysight N5746A
@@ -84,719 +81,6 @@ namespace DevicesControlLibrary.Devices.PowerSupply
         #endregion
 
         #region Public methods
-
-        #region Common commands
-
-        /// <summary>
-        ///     This command causes the following actions on the status system:
-        /// <list type="bullet">
-        ///     <item>
-        ///         <description>
-        ///             Clears the Standard Event Status, Operation Status Event, 
-        ///             and Questionable Status Event registers
-        ///         </description>
-        ///     </item>
-        /// <item>
-        ///         <description>
-        ///             Clears the Status Byte and the Error Queue
-        ///         </description>
-        ///     </item>
-        /// <item>
-        ///         <description>
-        ///             If *CLS immediately follows a program message terminator(NL), 
-        ///             then the output queue and the MAV bit are also cleared.
-        ///         </description>
-        ///     </item>
-        /// </list>
-        /// </summary>
-        public void SetClearStatus()
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("*CLS;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set clear status command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     All of the enabled events of the Standard Event Status Event Register 
-        ///     are logically OR-ed to cause the Event Summary Bit(ESB) of the Status
-        ///     Byte Register to be set.The query reads the Standard Event The query
-        ///     reads the Standard Event Status Enable register. The bit configuration of 
-        ///     the Standard Event register is as <see cref="SetStandartEventStatusEnable(int)">follows</see>.
-        /// </summary>
-        /// <returns>Int value</returns>
-        public int GetStandartEventStatus()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestInt("*ESE?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to get result standart event status enabled command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command programs the Standard Event Status Enable register bits. 
-        ///     The programming determines which events of the Standard Event Status 
-        ///     Event register (see *ESR?) are allowed to set the ESB (Event Summary Bit)
-        ///      of the Status Byte register. A "1" in the bit position enables the corresponding event. 
-        /// </summary>
-        /// <param name="nrf">register of activation of the standard event status</param>
-        /// <remarks>
-        /// <list type="bullet">
-        ///     <item>
-        ///         <description>Bit position 7. Bit value 128. Bit name - PON | Power-on has occured</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>Bit position 6. Undefined</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>Bit position 5. Bit value 32. Bit name - CME | Command error</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>Bit position 4. Bit value 16. Bit name - EXE | Execution error</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>Bit position 3. Bit value 8. Bit name - DDE | Device-dependent error</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>Bit position 2. Bit value 4. Bit name - QUE | Query error</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>Bit position 1. Undefined</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>Bit position 0. Bit value 1. Bit name - OPC | Operation complete</description>
-        ///     </item>
-        /// </list>
-        /// </remarks>
-        public void SetStandartEventStatusEnable(int nrf)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("*ESE " + nrf + ";");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set standart event status enabled command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This query reads the Standard Event Status Event register. 
-        ///     Reading the register clears it. 
-        ///     The bit configuration is the same as the Standard Event Status Enable register(see <see cref="SetStandartEventStatusEnable(int)">*ESE</see>).
-        /// </summary>
-        /// <returns>Return int value</returns>
-        public int SetStandartEventStatusEnable()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestInt("*ESR?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to get result event status register command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This query requests the power supply to identify itself.
-        ///     It returns a string of four fields separated by commas.
-        /// </summary>
-        /// <returns>String, contains identification information</returns>
-        public string GetIdentificationInformation()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestString("*IDN?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to get result identification device command. Reason:  " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command causes the instrument to set the OPC bit (bit 0) of the
-        ///     Standard Event Status register when the instrument has completed all
-        ///     pending operations.
-        ///     Pending operations are complete when:
-        ///     <list type="bullet">
-        ///     <item>
-        ///         <description>All commands sent before *OPC have been executed.
-        ///         This includes overlapped commands.
-        ///         Most commands are sequential and are completed before the next command
-        ///         is executed.
-        ///         Commands that affect output voltage, current or state, relays, 
-        ///         and trigger actions are executed in parallel (or overlapped) with subsequent commands 
-        ///         sent to the power supply.
-        ///         The *OPC command provides notification that all overlapped commands have been completed.
-        ///         </description>
-        ///     </item>
-        ///     <item>
-        ///         <descriprion>All triggered actions are completed *OPC does not prevent processing
-        ///         of subsequent commands, but bit 0 will not be set until all pending operations are completed.
-        ///         *OPC? causes the instrument to place an ASCII "1" in the Output Queue when all pending operations
-        ///         are completed.Unlike *OPC, *OPC? prevents processing of all subsequent commands. 
-        ///         It can be used at the end of a command line so that the program can monitor the bus for data
-        ///         until it receives the "1" from the Output Queue.
-        ///         </descriprion>
-        ///     </item>
-        /// </list>
-        /// </summary>
-        public void SetOpcEnable()
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("*OPC;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set enable operation complete bit in ESR command. Reason:  " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command places a 1 in the Output Queue when all pending operations have completed.
-        ///     Because it requires your program to read the returned value before executing the next 
-        ///     program statement, *OPC? can be used to cause the controller to wait for commands to 
-        ///     complete before proceeding with its program.
-        /// </summary>
-        /// <returns>Return true if all comands is complete. Otherwise - return false</returns>
-        public bool GetOpc()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestInt("*OPC?;") == 1;
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to get result operation complete command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This query requests the unit to identify any installed options. A 0 indicates no options are installed.
-        /// </summary>
-        /// <returns>Return value of option number</returns>
-        /// <remarks>If return 0 - standart instrument, without option</remarks>
-        public int GetOpt()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestInt("*OPT?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception(
-                    "Failed to get result option number installed command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command restores the power supply to a state that was previously
-        ///     stored in memory locations 0 through 15 with the <see cref="SetSaveState">*SAV</see> command.
-        ///     Note that you can only recall a state from a location that contains a previously-stored state. 
-        /// </summary>
-        /// <param name="state">Value of memory location</param>
-        /// <remarks>state parametr value of range 0 to 15. Attention! All saved instrument states are lost when the unit is turned off.</remarks>
-        public void SetRecallSavedState(int state)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("*OPT?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception(
-                    "Failed to get result option number installed command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command resets the power supply to a factory-defined state. 
-        ///     This state is defined as follows.Note that *RST also forces an ABORt command. 
-        ///     The *RST settings are as follows: 
-        ///     <list type="bullet">
-        ///     <item>
-        ///         <description>CAL:STAT Off</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>INIT:CONT Off</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>OUTP Off</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>[SOUR:]CURR 0</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>[SOUR:]:CURR:TRIG 0</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>[SOUR:]CURR:PROT:STAT Off</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>[SOUR:]VOLT 0</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>[SOUR:]VOLT:LIM 0</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>[SOUR:]VOLT:TRIG 0</description>
-        ///     </item>
-        ///     <item>
-        ///         <description>[SOUR:]VOLT:PROT MAX</description>
-        ///     </item>
-        ///     </list>
-        /// </summary>
-        public void Reset()
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("*RST;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception(
-                    "Failed to reset command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command stores the present state of the power supply to memory locations 0 through 15. 
-        /// </summary>
-        /// <param name="state">Value of state number to range 0-15</param>
-        /// <remarks>All saved instrument states are lost when the unit is turned off.</remarks>
-        public void SetSaveState(int state)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("*SAV " + state + ";");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception(
-                    "Failed to set saved state command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command sets the condition of the Service Request Enable Register.
-        ///     This register determines which bits from the Status Byte Register are
-        ///     allowed to set the Master Status Summary(MSS) bit and the Request 
-        ///     for Service(RQS) summary bit. 
-        ///     A 1 in any Service Request Enable Register bit position enables the 
-        ///     corresponding Status Byte Register bit and all such enabled bits then
-        ///     are logically OR-ed to cause Bit 6 of the Status Byte Register to be set.
-        /// </summary>
-        /// <param name="nrf">Value of bit</param>
-        public void SetServiceRequestEnableRegister(int nrf)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("*SRE " + nrf + ";");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception(
-                    "Failed to set service request enable register command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     When the controller conducts a serial poll in response to SRQ, 
-        ///     the RQS bit is cleared, but the MSS bit is not.
-        ///     When *SRE is cleared (by programming it with 0), the power 
-        ///     supply cannot generate an SRQ to the controller.
-        ///     The query returns the current state of *SRE
-        /// </summary>
-        /// <returns>String value of SRE</returns>
-        public string GetServiceRequestEnableRegister()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestString("*SRE?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception(
-                    "Failed to get service request enable register command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This query reads the Status Byte register, which contains the status    
-        ///     summary bits and the Output Queue MAV bit. 
-        ///     Reading the Status Byte register does not clear it. 
-        ///     The input summary bits are cleared when the appropriate event registers are read.
-        ///     The MAV bit is cleared at poweron, by <see cref="SetClearStatus">*CLS</see> or
-        ///     when there is no more response data available.
-        ///     A serial poll also returns the value of the Status Byte register, 
-        ///     except that bit 6 returns Request for Service(RQS) instead of Master Status Summary(MSS).
-        ///     A serial poll clears RQS, but not MSS.When MSS is set, it indicates that the power supply 
-        ///     has one or more reasons for requesting service.
-        /// </summary>
-        /// <returns>
-        ///     String value of status byte:
-        ///     <list type="bullet">
-        ///         <item>
-        ///             <description>Bit position 7. Bit value 128. Bit name OPER | Operation status summary</description>
-        ///         </item>
-        ///         <item>
-        ///             <description>Bit position 6. Bit value 64. Bit name MSS (RQS) | Master status summary (Request for service)</description>
-        ///         </item>
-        ///         <item>
-        ///             <description>Bit position 5. Bit value 32. Bit name ESB | Event status byte summary</description>
-        ///         </item>
-        ///         <item>
-        ///             <description>Bit position 4. Bit value 16. Bit name MAV | Message aviable</description>
-        ///         </item>
-        ///         <item>
-        ///             <description>Bit position 3. Bit value 8. Bit name QUES | Questionable status summary</description>
-        ///         </item>
-        ///         <item>
-        ///             <description>Bit position 2. Bit value 4 Bit name ERR | Error queue not empty</description>
-        ///         </item>
-        ///         <item>
-        ///             <description>Bit position 1-0 undefined</description>
-        ///         </item>
-        ///     </list>
-        /// </returns>
-        public string GetStatusByte()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestString("*STB?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception(
-                    "Failed to get status byte register value command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command generates a trigger when the trigger source is set to BUS.
-        ///     The command has the same affect as the Group Execute Trigger(GET) command.
-        /// </summary>
-        public void SetTrigger()
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("*TRG;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception(
-                    "Failed to set trigger command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     Self-Test query
-        ///     TODO: ErrorCode translate
-        /// </summary>
-        public bool Test()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestString("*TST?;").Equals("0");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Could not process test command: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command instructs the power supply not to process any further
-        ///     commands until all pending operations are completed.
-        ///     Pending operations are as defined under the *OPC command. 
-        ///     *WAI can be aborted only by sending the power supply a Device Clear command.
-        /// </summary>
-        public void Wait()
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("*WAI;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Could not process wait command: " + exception.Message);
-            }
-        }
-
-        #endregion
-
-        #region Calibration comands
-
-        /// <summary>
-        ///     This command initiates the calibration of the output current. 
-        /// </summary>
-        /// <param name="maximumValue">The maximum current in the range of values for the output signal specified during calibration</param>
-        public void SetCalibrationOutputCurrent(double maximumValue)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("CAL:CURR " + maximumValue + ";");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set calibration in value of " + maximumValue + "A command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command enters a calibration value that you obtain by reading an external meter.
-        ///     You must first select a calibration level (with CALibrate:LEVel) for the value being entered.
-        ///     Data values are entered in either volts or amperes, depending on which function is being calibrated.
-        /// </summary>
-        /// <param name="dataValue">Value of calibration</param>
-        public void SetCalibrationData(double dataValue)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("CAL:DATA " + dataValue + ";");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set calibration data in value of " + dataValue + " command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command stores the date the unit was last calibrated. 
-        ///     The data must be of the numeric format “yyyy/mm/dd” 
-        ///     where yyyy indicates the year, mm indicates the month,
-        ///     and dd indicates the day.    
-        /// </summary>
-        /// <param name="date">Value of date in format yyyy/mm/dd</param>
-        public void SetCalibrationDate(string date)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("CAL:DATE " + date + ";");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set calibration date in value of " + date + " command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     The query returns the date.
-        /// </summary>
-        /// <returns>String, contains date in format yyyy/mm/dd</returns>
-        public string GetCalibrationDate()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestString("CAL:DATE?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to get calibration date value of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-
-        /// <summary>
-        ///     This command selects the next point in the calibration sequence.
-        /// </summary>
-        /// <param name="calibrationLevel">Value of calibration level:
-        /// <list type="bullet">
-        ///     <item>
-        ///         <description>
-        ///             P1 is the first calibration point
-        ///     </description>
-        ///     </item>
-        /// 
-        ///     <item>
-        ///         <description>
-        ///             P2 is the second calibration point
-        ///     </description>
-        ///     </item>
-        /// </list>
-        /// </param>
-        public void SetCalibrationLevel(CalibrationLevel calibrationLevel)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("CAL:LEV " + calibrationLevel + ";");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set calibration level value of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///      This command returns the next point in the calibration sequence.
-        /// </summary>
-        /// <returns>Value of calibration level:
-        /// <list type="bullet">
-        ///     <item>
-        ///         <description>
-        ///             P1 is the first calibration point
-        ///     </description>
-        ///     </item>
-        /// 
-        ///     <item>
-        ///         <description>
-        ///             P2 is the second calibration point
-        ///     </description>
-        ///     </item>
-        /// </list></returns>
-        public string GetCalibrationLevel()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestString("CAL:LEV?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to get calibration level value of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command lets you change the calibration password.
-        ///     A new password is automatically stored in nonvolatile memory.
-        ///     If the password is set to 0, password protection is removed and 
-        ///     the ability to enter calibration mode is unrestricted.The default
-        ///     password is 0 (zero).
-        /// </summary>
-        /// <param name="password">Password can contains only integer value</param>
-        public void SetCalibrationPassword(string password)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("CAL:PASS " + password + ";");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set calibration password value of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command enables/disables calibration mode. 
-        ///     Calibration mode must be enabled for the power supply 
-        ///     to accept any other calibration commands.
-        ///     The first parameter specifies the enabled or disabled state On(1) or Off(0).
-        ///     The second parameter is the password.
-        /// </summary>
-        /// <param name="state">True - state is on, False - state is off</param>
-        /// <param name="password">Password value can contains only int values. Optional</param>
-        public void SetCalibrationState(bool state, string password = null)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("CAL:STAT " +
-                                                 (state
-                                                     ? (password != null ? "ON [," + password + "]" : "ON")
-                                                     : "OFF") + ";");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set state calibration in value of " +
-                                    (state ? (password != null ? "ON [," + password + "]" : "ON") : "OFF") +
-                                    " command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     The query returns only the state, not the password.
-        /// </summary>
-        /// <returns>True - state is on, False - state is off</returns>
-        public bool GetCalibrationState()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestInt("CAL:STAT?;") == 1;
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to get state calibration value of command. Reason: " + exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command initiates the calibration of the output voltage. 
-        /// </summary>
-        /// <param name="voltageValueOfCalibration">Value of voltage</param>
-        public void SetCalibrationVoltage(double voltageValueOfCalibration)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("CAL:VOLT " + voltageValueOfCalibration + ";");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception(
-                    "Failed to set initiates the calibration of the output voltage value of command. Reason: " +
-                    exception.Message);
-            }
-        }
-
-        #endregion
-
-        #region Measure commands
-
-        /// <summary>
-        ///     Returns current consumption in amperes
-        /// </summary>
-        /// <return>Value of current in ampers</return>
-        public double GetMeasureCurrent()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestDouble("MEAS:CURR?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to get measurement current in amperes value of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     Returns output voltage in volts
-        /// </summary>
-        /// <return>Value of output voltage in volts</return>
-        public double GetMeasureVolt()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestDouble("MEAS:VOLT?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to get measurement output voltage in volt value of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        #endregion
 
         #region Output commands
 
@@ -1770,169 +1054,6 @@ namespace DevicesControlLibrary.Devices.PowerSupply
                                     exception.Message);
             }
         }
-
-        #endregion
-
-        #region Trigger commands
-
-        #region Abort commands
-
-        /// <summary>
-        ///     This command cancels any trigger actions in progress and returns the
-        ///     trigger system to the IDLE state, unless INIT:CONT is enabled.
-        ///     It also resets the WTG bit in the Status Operation Condition register.
-        ///     ABORt is executed at power-on and upon execution of* RST.
-        /// </summary>
-        public void SetTriggerAbort()
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("ABOR;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set the abort of command of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        #endregion
-
-        #region Initiate commands
-
-        /// <summary>
-        ///     This command controls the enabling of output triggers.
-        ///     When a trigger is enabled, a trigger causes the specified triggering action to occur.
-        ///     If the trigger system is not enabled, all triggers are ignored.
-        /// </summary>
-        public void SetInitiate()
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("INIT;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set the enabling of output triggers of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command controls the enabling of output triggers.
-        ///     When a trigger is enabled, a trigger causes the specified triggering action to occur.
-        ///     If the trigger system is not enabled, all triggers are ignored.
-        /// </summary>
-        /// <param name="state">True - is on, False - is off</param>
-        public void SetInitiateContinuous(bool state)
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("INIT:CONT " + (state ? "ON" : "OFF") + ";");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set the enabling of output triggers of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command continuously initiates output triggers.
-        ///     The enabled state is On(1); the disabled state is Off(0).
-        ///     When disabled, the trigger system must be initiated for each 
-        ///     trigger with the INITiate command.
-        /// </summary>
-        /// <returns>True - is on, False - is off</returns>
-        public bool GetInitiateContinuous()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestInt("INIT:CONT?;") == 1;
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to get the enabling of output triggers of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        #endregion
-
-        #region Trigger commands
-
-        /// <summary>
-        ///     If the trigger system has been initiated, this command generates an 
-        ///     immediate output trigger.
-        ///     When sent, the output trigger will:
-        ///     - Initiate an output change as specified by the CURR:TRIG or VOLT:TRIG settings.
-        ///     - Clear the WTG bits in the Status Operation Condition register after the trigger action has completed.
-        /// </summary>
-        public void SetTriggerImmediate()
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("TRIG;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set trigger immediate of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command selects the trigger source for the output trigger system.
-        ///     Only BUS can be selected as the trigger source.
-        /// </summary>
-        public void SetTriggerSource()
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("TRIG:SOUR BUS;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set trigger source in BUS of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     Get trigger source
-        /// </summary>
-        /// <returns>Value of trigger source</returns>
-        public string GetTriggerSource()
-        {
-            try
-            {
-                return _lanExchanger.SendWithRequestString("TRIG:SOUR?;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to get trigger source of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        /// <summary>
-        ///     This command generates a trigger when the trigger source is set to BUS.
-        ///     The command has the same affect as the Group Execute Trigger(GET) command.
-        /// </summary>
-        public void SetTriggerGenerated()
-        {
-            try
-            {
-                _lanExchanger.SendWithoutRequest("TRG*;");
-            }
-            catch (Exception exception)
-            {
-                throw new Exception("Failed to set generated trigger of command. Reason: " +
-                                    exception.Message);
-            }
-        }
-
-        #endregion
 
         #endregion
