+++ Devices/PowerSupply/Subsystems/Measure/ISubsystemMeasure.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DevicesControlLibrary.Devices.PowerSupply.Subsystems.Measure
+{
+    public interface ISubsystemMeasure
+    {
+        double GetMeasureCurrent();
+        double GetMeasureVolt();
+    }
+}
