+++ Devices/PowerSupply/Subsystems/Source/ISubsystemSource.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DevicesControlLibrary.Devices.PowerSupply.Subsystems.Source
+{
+    public interface ISubsystemSource
+    {
+    }
+}
