+++ Devices/PowerSupply/Subsystems/Status/ISubsystemStatus.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DevicesControlLibrary.Devices.PowerSupply.Subsystems.Status
+{
+    public interface ISubsystemStatus
+    {
+    }
+}
